+++ src/eca/handlers.clj
@@ -7,14 +7,14 @@
    [eca.logger :as logger]))
 
 (defn ^:private initialize-extra-models! [db*]
-  (let [config (config/all)]
+  (let [config (config/all @db*)]
     (when-let [ollama-models (seq (llm-api/extra-models config))]
       (swap! db* update :models concat (map #(str config/ollama-model-prefix (:model %)) ollama-models)))))
 
 (defn initialize [{:keys [db*]} params]
   (logger/logging-task
    :eca/initialize
-   (let [config (config/all)]
+   (let [config (config/all @db*)]
      (swap! db* assoc
             :client-info (:client-info params)
             :workspace-folders (:workspace-folders params)
@@ -36,7 +36,7 @@
 (defn chat-prompt [{:keys [messenger db*]} params]
   (logger/logging-task
    :eca/chat-prompt
-   (let [config (config/all)]
+   (let [config (config/all @db*)]
      (f.chat/prompt params db* messenger config))))
 
 (defn chat-query-context [{:keys [db*]} params]